Update gemspec with TRX contact info
In code review it was suggested that (correctly) we should include
contact information for our forked version.

I'm not sure what the etiquette is but it feels like wiping out the
original author's credit would be harsh, even if they're no longer
maintaining it.

Have also included a link to the original in the description. I think
mentioning the fork which we forked from may be going a bit too far.